requirements.txt
joblib==1.1.1
JPype1==1.2.1
pandas==1.1.5
python-dateutil==2.8.2
pytorch-crf==0.7.2
pytorch-pretrained-bert==0.6.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
torch==1.9.0
transformers==4.25.1
zipp==3.6.0